server/requirements.txt
Django==2.2.17
flake8==3.8.4
mccabe==0.6.1
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pyflakes==2.2.0
pytz==2020.5
sqlparse==0.4.1
djangorestframework==3.12.2
django-cors-headers==3.6.0
django-filter==2.4.0
